Ignore text fixtures

+++ app/controllers/tasks_controller.rb
@@ -1,17 +1,21 @@
 class TasksController < ApplicationController
-  before_action :load_task, only: %(show update destroy)
+  before_action :authenticate_user_using_x_auth_token
+
+  before_action :load_task, only: %i[show update destroy]
 
   def index
-    tasks = Task.all
-    render status: :ok, json: { tasks: tasks }
+    tasks = policy_scope(Task)
+    render status: :ok, json: { tasks: { pending: tasks.organize(:pending), completed: tasks.organize(:completed) } }
   end
 
   def show
-    render status: :ok, json: { task: @task }
+    comments = @task.comments.order('created_at DESC')
+    render status: :ok, json: { task: @task, assigned_user: @task.user, comments: comments }
   end
 
   def create
-    @task = Task.new(task_params)
+    @task = Task.new(task_params.merge(creator_id: @current_user.id))
+    authorize @task
     if @task.save
       render status: :ok, json: { notice: 'Task was successfully created' }
     else
@@ -21,6 +25,7 @@ class TasksController < ApplicationController
   end
 
   def update
+    authorize @task
     if @task.update(task_params)
       render status: :ok, json: { notice: 'Successfully updated task.' }
     else
@@ -29,6 +34,7 @@ class TasksController < ApplicationController
   end
 
   def destroy
+    authorize @task
     if @task.destroy
       render status: :ok, json: { notice: 'Successfully deleted task.' }
     else
@@ -39,12 +45,12 @@ class TasksController < ApplicationController
   private
 
   def task_params
-    params.require(:task).permit(:title, :user_id)
+    params.require(:task).permit(:title, :user_id, :progress, :status)
   end
 
   def load_task
     @task = Task.find(params[:id])
-  rescue ActiveRecord::RecordNotFound => e
+    rescue ActiveRecord::RecordNotFound => e
     render json: { errors: e }
   end
 end
